refactor(dashboard): move metrics logic from events.service to dashboard.service
