+++ frontend/src/services/events.service.js
@@ -59,12 +59,3 @@ export const getFavoritesReport = () =>
     }).then(r => r.json())
 
 
-export const getDashboardMetrics = () =>
-    fetch(`${API_BASE}/dashboard/metrics`, {
-        headers: getHeaders(),
-    }).then(r => r.json())
-
-export const getSalesByEvent = () =>
-    fetch(`${API_BASE}/dashboard/sales-by-event`, {
-        headers: getHeaders(),
-    }).then(r => r.json())
