+++ frontend/src/services/index.js
@@ -7,4 +7,5 @@ export * from './events.service'
 export * from './categories.service'
 export * from './ticketTypes.service'
 export * from './checkout.service'
+export * from './dashboard.service'
 export { API_BASE } from './config'
